# server/mcp/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "evcc",
    "description": "Solar charging. Super simple.",
    "version": "0.2.0",
    "contact": {
      "url": "https://github.com/evcc-io/evcc"
    }
  },
...
  "paths": {
...
    "/session/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/id"
        }
      ],
      "put": {
        "operationId": "updateSession",
        "summary": "Update charging session",
        "description": "Update vehicle, loadpoint or odometer of a charging session. Only provided fields are changed; a null odometer clears the stored value.",
        "externalDocs": {
          "url": "https://docs.evcc.io/en/features/sessions"
        },
        "tags": [
          "sessions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "vehicle": {
                    "$ref": "#/components/schemas/VehicleName"
                  },
                  "loadpoint": {
                    "$ref": "#/components/schemas/LoadpointName"
                  },
                  "odometer": {
                    "$ref": "#/components/schemas/Odometer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "operationId": "deleteSession",
        "summary": "Delete charging session",
        "description": "Delete charging session.",
        "externalDocs": {
          "url": "https://docs.evcc.io/en/features/sessions"
        },
        "tags": [
          "sessions"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/NullResult"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Id": {
        "description": "Loadpoint index starting at 1.",
        "type": "integer",
        "example": 1,
        "minimum": 1
      },
      "LoadpointName": {
        "externalDocs": {
          "url": "https://docs.evcc.io/en/reference/configuration/loadpoints#title"
        },
        "type": "string",
        "example": "Garage"
      },
...
      "Odometer": {
        "nullable": true,
        "type": "number",
        "description": "Vehicle odometer reading in kilometers"
      },
...
      "VehicleName": {
        "externalDocs": {
          "url": "https://docs.evcc.io/en/reference/configuration/vehicles#name"
        },
        "type": "string",
        "minLength": 1,
        "pattern": "[a-zA-Z0-9_.:-]+",
        "example": "vehicle_1"
      },
...
    },
    "parameters": {
...
      "id": {
        "name": "id",
        "description": "Loadpoint index starting at 1",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/Id"
        }
      },
...
    },
    "responses": {
...
      "NullResult": {
        "description": "Success - Null result",
        "content": {
          "application/json": {
            "schema": {
              "description": "Value is always null",
              "nullable": true,
              "type": "object"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Long-lived API key (prefixed `evcc_`, managed via /auth/apikey).\nGrants access to /api/db/* without requiring the X-Admin-Password header.\n"
      }
    }
  }
}